ISSUE #3829 move other files into models folder
