+++ backend/tests/v5/e2e/routes/teamspaces/projects/models/federations/revisions.test.js
@@ -16,8 +16,8 @@
  */
 
 const SuperTest = require('supertest');
-const ServiceHelper = require('../../../../../helper/services');
-const { src } = require('../../../../../helper/path');
+const ServiceHelper = require('../../../../../../helper/services');
+const { src } = require('../../../../../../helper/path');
 
 const { templates } = require(`${src}/utils/responseCodes`);
